[circleci] Inline scrape_warning_messages
This job was only being used for process_artifacts_combined in circleci
but as it was a separate job would incur the cost and time of spinning
up a new CI worker. We can simply inline this into
process_artifacts_combined.

Test plan: Download
[build.tgz](https://app.circleci.com/pipelines/github/facebook/react/57183/workflows/3b22a99e-b8a5-4c80-9682-576f24581b00/jobs/956476/artifacts)
from circleci and observe that the __test_utils__ directory is still
present.

ghstack-source-id: 8748d8fb9cf01d8fffe7e7590729a44def265c79
Pull Request resolved: https://github.com/facebook/react/pull/30388