[circleci] Remove build_and_test workflow
Now that we've fully migrated the PR CI workflow (build_and_test) to GH,
we can fully delete the workflow altogether. This will break
https://react-builds.vercel.app/ unfortunately but I'll wait for
acdlite to come back from vacation and work with him to fix it.

The remaining jobs in circleci are for publishing prereleases of React
to npm. I'll work on migrating those next.

ghstack-source-id: 05f346829f47970592ce855340e02b37285635db
Pull Request resolved: https://github.com/facebook/react/pull/30480